requirements.txt
aiofiles==23.2.1
altair==5.1.2
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
av==10.0.0
blis==0.7.11
Brotli==1.1.0
catalogue==2.0.10
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.7.0
comm==0.1.4
confection==0.1.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
docker-pycreds==0.4.0
docopt==0.6.2
dora-search==0.1.12
einops==0.7.0
encodec==0.1.1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.9.0
flashy==0.0.2
fonttools==4.44.0
fsspec==2023.10.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.17.3
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
julius==0.2.7
jupyter_client==8.5.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lameenc==1.6.3
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
lightning-utilities==0.9.0
llvmlite==0.41.1
MarkupSafe==2.1.2
matplotlib==3.8.1
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
murmurhash==1.0.10
mutagen==1.47.0
nest-asyncio==1.5.8
networkx==3.0
num2words==0.5.13
numba==0.58.1
numpy==1.24.1
omegaconf==2.3.0
openunmix==1.2.1
orjson==3.9.10
packaging==23.2
pandas==2.1.2
parso==0.8.3
pathy==0.10.3
Pillow==9.3.0
platformdirs==3.11.0
pooch==1.8.0
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.19.0
pydantic==1.10.13
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
#pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.28.1
retrying==1.3.4
rpds-py==0.12.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spotipy==2.23.0
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
submitit==1.5.0
sympy==1.12
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.14.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
treetable==0.2.5
typer==0.7.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.24.0
wandb==0.16.0
wasabi==1.1.2
wcwidth==0.2.9
websockets==11.0.3
xformers==0.0.22.post7
yt-dlp==2023.10.13
zipp==3.17.0
